refactor(cli): move self-update under 'self' group
