+++ cmd/self/self.go
@@ -7,8 +7,10 @@ import (
 
 // NewSelfCmd builds the 'devsy self' parent command for managing the devsy CLI itself.
 func NewSelfCmd(globalFlags *flags.GlobalFlags) *cobra.Command {
-	return &cobra.Command{
+	cmd := &cobra.Command{
 		Use:   "self",
 		Short: "Manage the devsy CLI itself",
 	}
+	cmd.AddCommand(NewUpdateCmd())
+	return cmd
 }
