+++ cmd/self/update_test.go
@@ -1,4 +1,4 @@
-package cmd
+package self
 
 import (
 	"testing"
@@ -7,34 +7,34 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestNewSelfUpdateCmd_CommandName(t *testing.T) {
-	cmd := NewSelfUpdateCmd()
-	assert.Equal(t, "self-update", cmd.Use)
+func TestNewUpdateCmd_CommandName(t *testing.T) {
+	cmd := NewUpdateCmd()
+	assert.Equal(t, "update", cmd.Use)
 }
 
-func TestNewSelfUpdateCmd_HasVersionFlag(t *testing.T) {
-	cmd := NewSelfUpdateCmd()
+func TestNewUpdateCmd_HasVersionFlag(t *testing.T) {
+	cmd := NewUpdateCmd()
 	f := cmd.Flags().Lookup("version")
 	require.NotNil(t, f, "--version flag must exist")
 	assert.Equal(t, "", f.DefValue)
 }
 
-func TestNewSelfUpdateCmd_HasDryRunFlag(t *testing.T) {
-	cmd := NewSelfUpdateCmd()
+func TestNewUpdateCmd_HasDryRunFlag(t *testing.T) {
+	cmd := NewUpdateCmd()
 	f := cmd.Flags().Lookup("dry-run")
 	require.NotNil(t, f, "--dry-run flag must exist")
 	assert.Equal(t, "false", f.DefValue)
 }
 
-func TestNewSelfUpdateCmd_HasChannelFlag(t *testing.T) {
-	cmd := NewSelfUpdateCmd()
+func TestNewUpdateCmd_HasChannelFlag(t *testing.T) {
+	cmd := NewUpdateCmd()
 	f := cmd.Flags().Lookup("channel")
 	require.NotNil(t, f, "--channel flag must exist")
 	assert.Equal(t, channelStable, f.DefValue)
 }
 
-func TestNewSelfUpdateCmd_AcceptsNoPositionalArgs(t *testing.T) {
-	cmd := NewSelfUpdateCmd()
+func TestNewUpdateCmd_AcceptsNoPositionalArgs(t *testing.T) {
+	cmd := NewUpdateCmd()
 	err := cmd.Args(cmd, []string{"unexpected"})
 	assert.Error(t, err, "should reject positional arguments")
 
@@ -42,17 +42,17 @@ func TestNewSelfUpdateCmd_AcceptsNoPositionalArgs(t *testing.T) {
 	assert.NoError(t, err, "should accept zero arguments")
 }
 
-func TestNewSelfUpdateCmd_RejectsInvalidChannel(t *testing.T) {
-	cmd := NewSelfUpdateCmd()
+func TestNewUpdateCmd_RejectsInvalidChannel(t *testing.T) {
+	cmd := NewUpdateCmd()
 	cmd.SetArgs([]string{"--channel", "nightly"})
 	err := cmd.Execute()
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "invalid channel")
 }
 
-func TestNewSelfUpdateCmd_AcceptsValidChannels(t *testing.T) {
+func TestNewUpdateCmd_AcceptsValidChannels(t *testing.T) {
 	for _, ch := range []string{channelStable, channelBeta} {
-		cmd := NewSelfUpdateCmd()
+		cmd := NewUpdateCmd()
 		require.NoError(t, cmd.Flags().Set("channel", ch))
 		preRunE := cmd.PreRunE
 		require.NotNil(t, preRunE)
