+++ e2e/e2e_suite_test.go
@@ -27,7 +27,7 @@ import (
 	_ "github.com/devsy-org/devsy/e2e/tests/provider"
 	_ "github.com/devsy-org/devsy/e2e/tests/rename"
 	_ "github.com/devsy-org/devsy/e2e/tests/runusercommands"
-	_ "github.com/devsy-org/devsy/e2e/tests/selfupdate"
+	_ "github.com/devsy-org/devsy/e2e/tests/self"
 	_ "github.com/devsy-org/devsy/e2e/tests/ssh"
 	_ "github.com/devsy-org/devsy/e2e/tests/template"
 	_ "github.com/devsy-org/devsy/e2e/tests/tunnel"
